requirements.txt
Django==5.1.5
djangorestframework==3.15.0
whitenoise==6.6.0
gunicorn==21.2.0 